Indicator UI Arm, Leg bug fix

+++ Assets/Scripts/Bullet.cs
@@ -54,37 +54,37 @@ public class Bullet : MonoBehaviourPunCallbacks
 
             if (collision.collider.CompareTag("PlayerHead"))
             {
-                collision.transform.root.gameObject.GetPhotonView().RPC("GetDamage", RpcTarget.OthersBuffered, 100f);
+                collision.transform.root.gameObject.GetPhotonView().RPC("GetDamage", RpcTarget.OthersBuffered, (int)100f);
                 SniperGameManager.Instance.GetUI.UpdateIndicator("Head", 2);
                 SniperGameManager.Instance.GetUI.gameObject.GetPhotonView().RPC("UpdateIndicator", RpcTarget.OthersBuffered, "Head", 1);
             }
             else if (collision.collider.CompareTag("PlayerBody"))
             {
-                collision.transform.root.gameObject.GetPhotonView().RPC("GetDamage", RpcTarget.OthersBuffered, Random.Range(70f, 99f));
+                collision.transform.root.gameObject.GetPhotonView().RPC("GetDamage", RpcTarget.OthersBuffered, (int)Random.Range(70f, 99f));
                 SniperGameManager.Instance.GetUI.UpdateIndicator("Body", 2);
                 SniperGameManager.Instance.GetUI.gameObject.GetPhotonView().RPC("UpdateIndicator", RpcTarget.OthersBuffered, "Body", 1);
             }
             else if (collision.collider.CompareTag("PlayerArmL"))
             {
-                collision.transform.root.gameObject.GetPhotonView().RPC("GetDamage", RpcTarget.OthersBuffered, Random.Range(10f, 50f));
+                collision.transform.root.gameObject.GetPhotonView().RPC("GetDamage", RpcTarget.OthersBuffered, (int)Random.Range(10f, 50f));
                 SniperGameManager.Instance.GetUI.UpdateIndicator("ArmL", 2);
                 SniperGameManager.Instance.GetUI.gameObject.GetPhotonView().RPC("UpdateIndicator", RpcTarget.OthersBuffered, "ArmL", 1);
             }
             else if (collision.collider.CompareTag("PlayerArmR"))
             {
-                collision.transform.root.gameObject.GetPhotonView().RPC("GetDamage", RpcTarget.OthersBuffered, Random.Range(10f, 50f));
+                collision.transform.root.gameObject.GetPhotonView().RPC("GetDamage", RpcTarget.OthersBuffered, (int)Random.Range(10f, 50f));
                 SniperGameManager.Instance.GetUI.UpdateIndicator("ArmR", 2);
                 SniperGameManager.Instance.GetUI.gameObject.GetPhotonView().RPC("UpdateIndicator", RpcTarget.OthersBuffered, "ArmR", 1);
             }
             else if (collision.collider.CompareTag("PlayerLegL"))
             {
-                collision.transform.root.gameObject.GetPhotonView().RPC("GetDamage", RpcTarget.OthersBuffered, Random.Range(20f, 60f));
+                collision.transform.root.gameObject.GetPhotonView().RPC("GetDamage", RpcTarget.OthersBuffered, (int)Random.Range(20f, 60f));
                 SniperGameManager.Instance.GetUI.UpdateIndicator("LegL", 2);
                 SniperGameManager.Instance.GetUI.gameObject.GetPhotonView().RPC("UpdateIndicator", RpcTarget.OthersBuffered, "LegL", 1);
             }
             else if (collision.collider.CompareTag("PlayerLegR"))
             {
-                collision.transform.root.gameObject.GetPhotonView().RPC("GetDamage", RpcTarget.OthersBuffered, Random.Range(20f, 60f));
+                collision.transform.root.gameObject.GetPhotonView().RPC("GetDamage", RpcTarget.OthersBuffered, (int)Random.Range(20f, 60f));
                 SniperGameManager.Instance.GetUI.UpdateIndicator("LegR", 2);
                 SniperGameManager.Instance.GetUI.gameObject.GetPhotonView().RPC("UpdateIndicator", RpcTarget.OthersBuffered, "LegR", 1);
             }
